% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmatch}
\alias{fmatch}
\title{Fast string matching with limited wildcard support}
\usage{
fmatch(x, y)
}
\arguments{
\item{x}{character}

\item{y}{character}
}
\value{
integer positions of matches in \code{y} (\code{NA} indicates no match)
}
\description{
For each string in `x`, finds the _first_ match in _y_.
The character '?' is treated as a wildcard in both `x` and `y`.
}
