% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmatch}
\alias{fmatch}
\title{Fast string matching with limited wildcard support}
\usage{
fmatch(x, y)
}
\arguments{
\item{x}{character}

\item{y}{character}
}
\value{
integer positions of matches in \code{y} (\code{NA} indicates no match)
}
\description{
For each string in \code{x}, find the find match in \code{y}.
The character '?' is treated as a wildcard in both \code{x} and \code{y}.
}
